Updated install and uninstall scripts with improved error handling.

+++ utils/install_inetvis.sh
@@ -8,8 +8,20 @@ APPS_PATH=/usr/share/applications/
 USR_BIN=/usr/local/bin
 
 echo "Creating directories..."
-sudo mkdir /opt/${BIN_NAME}-${VER}
-sudo ln -s /opt/${BIN_NAME}-${VER} /opt/${BIN_NAME}
+
+OPT_DIR_VER=/opt/${BIN_NAME}-${VER}
+if [[ ! -d "${OPT_DIR_VER}" ]]; then
+    sudo mkdir ${OPT_DIR_VER}
+else
+    echo "Not creating ${OPT_DIR_VER} as it already exists"
+fi
+
+OPT_DIR_SYM=/opt/${BIN_NAME}
+if [[ ! -L "${OPT_DIR_SYM}" ]]; then
+    sudo ln -s /opt/${BIN_NAME}-${VER} /opt/${BIN_NAME}
+else
+    echo "Symlink for ${OPT_DIR_SYM} already exists. Please link manually to ${OPT_DIR_VER} if desired."
+fi
 
 echo "Copying files..."
 sudo cp -r bin/ /opt/${BIN_NAME}-${VER}
@@ -21,7 +33,12 @@ echo "Copying menu entry..."
 sudo cp resources/${BIN_NAME}.desktop ${APPS_PATH} 
 
 echo "Creating symlink for inetvis..."
-sudo ln -s /opt/${BIN_NAME}/bin/${BIN_NAME} ${USR_BIN} 
+
+if [[ ! -L "${USR_BIN}/${BIN_NAME}" ]]; then
+    sudo ln -s /opt/${BIN_NAME}/bin/${BIN_NAME} ${USR_BIN} 
+else
+    echo "Symlink at ${USR_BIN}/${BIN_NAME} already exists. Please update if necessary."
+fi
 
 echo "Setting capabilities on binary to allow monitoring interfaces"
 sudo setcap cap_net_raw,cap_net_admin=eip /opt/${BIN_NAME}/bin/${BIN_NAME}
